Support named agent options config groups in Maven plugin
This commit modifies the previous approach by introducing a named
`config` block that allows the user to have separate groups of agent
options that can be selected when executing maven from the command line.

The current is no default group. The "default" group below is merely an
example.

The access filter configuration for tests has also been moved to the
src/test/resources folder.

<configuration>
  <agentOptions>
    <config name="default">
      <agentOption>experimental-class-loader-support</agentOption>
    </config>
    <config name="test">
      <agentOption>experimental-class-loader-support</agentOption>
      <agentOption>access-filter-file=${basedir}/src/test/resources/access-filter.json</agentOption>
    </config>
  </agentOptions>
</configuration>

See #162